Added option to remove scrollbars from full screen
In the advanced options there is a new field, FixedPageUI ->
ShowScrollBarsInFullScreen which is set to true by default. Unless this
is set to false then this changeset will have no effect. If it is set to
false then when full screen view is entered, with a zoom level like fit
content which usually requires scroll bars, then the pages will occupy
the entire screen area rather than the screen area minus scroll bars and
adding scroll bars. This is useful because presentation view does not
support facing pages.